+++ R/ppc-distributions.R
@@ -188,8 +188,9 @@ ppc_freqpoly <- function(y, yrep, ...,
     facet_bg(FALSE)
 }
 
-#' @export
 #' @rdname PPC-distributions
+#' @export
+#' @template args-group
 #'
 ppc_freqpoly_grouped <-
   function(y,
@@ -345,7 +346,6 @@ ppc_ecdf_overlay <- function(y, yrep, ...,
 
 #' @export
 #' @rdname PPC-distributions
-#' @template args-group
 #' @param probs A numeric vector passed to \code{\link[ggplot2]{geom_violin}}'s
 #'   \code{draw_quantiles} argument to specify at which quantiles to draw
 #'   horizontal lines. Set to \code{NULL} to remove the lines.